mode-specific leader keys should only trigger on normal mode.
once again fixing that bug where SPC was triggering prefix key in
insert mode. It was totally working before, no idea what was going
on. I wouldn't be surprised if the bug comes back at this point... But
this fix makes sense so at the same time I'm thinking it might be
fixed for good.